Feat/#18 채팅 모듈 통합 및 기능 추가 + security 설정 수정

+++ backendProject/src/main/java/likelion/mlb/backendProject/domain/match/dto/RoundInfo.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 @AllArgsConstructor
 public class RoundInfo {
     private UUID id;
-    private short no;
+    private Integer no;
     private String openAt;
     private String lockAt;
 }
